refactor: remove test abstraction
